feat(api,mcp): persistence + media parity; peek counts to u32 (#195)
sram_get/sram_set, export_spc, screenshot {native, bg}, decode_sprites; peek_vram/peek_aram count u16->u32 (one-call full dumps). Closes #173.

+++ crates/luna-cli/src/state.rs
@@ -544,9 +544,9 @@ pub(crate) fn run_state(
             match parse_assert_spec_no_bank(spec) {
                 Ok((offset, want)) => {
                     let res = if kind == 0 {
-                        em.peek_aram(offset, want.len() as u16)
+                        em.peek_aram(offset, want.len() as u32)
                     } else {
-                        em.peek_vram(offset, want.len() as u16)
+                        em.peek_vram(offset, want.len() as u32)
                     };
                     match res {
                         Ok(got) if got == want => {
